requirements.txt
Flask==3.0.3
gunicorn==21.2.0
PyPDF2==3.0.1
reportlab==4.0.8
google-cloud-vision==3.4.4
google-auth==2.35.0
requests==2.31.0
oauthlib==3.2.2
resend==0.8.0
python-dotenv==1.0.0
certifi==2024.8.30
urllib3==2.0.7
selenium==4.25.0
